Put config generation one type per file
